Move canopy context out of index
Moves the CanopyContext definition out of index and into its own file.
This simplifies the index file to just be concerned with exporting
objects.
